actually use new exporters
