+++ R/s2-geography.R
@@ -136,22 +136,20 @@ as_s2_geography.logical <- function(x, ...) {
 #' @rdname as_s2_geography
 #' @export
 as_wkb.s2_geography <- function(x, ...) {
-  wk::new_wk_wkb(
-    s2_geography_to_wkb(x, wk::wk_platform_endian()),
-    crs = wk::wk_crs_longlat("WGS84"),
-    geodesic = TRUE
-  )
+  wkb <- wk::wk_handle(x, wk::wkb_writer())
+  wk::wk_is_geodesic(wkb) <- TRUE
+  wk::wk_crs(wkb) <- wk::wk_crs_longlat()
+  wkb
 }
 
 #' @importFrom wk as_wkt
 #' @rdname as_s2_geography
 #' @export
 as_wkt.s2_geography <- function(x, ...) {
-  wk::new_wk_wkt(
-    s2_geography_to_wkt(x, precision = 16, trim = TRUE),
-    crs = wk::wk_crs_longlat(),
-    geodesic = TRUE
-  )
+  wkt <- wk::wk_handle(x, wk::wkt_writer())
+  wk::wk_is_geodesic(wkt) <- TRUE
+  wk::wk_crs(wkt) <- wk::wk_crs_longlat()
+  wkt
 }
 
 
@@ -171,7 +169,7 @@ as_wkt.s2_geography <- function(x, ...) {
 
 #' @export
 format.s2_geography <- function(x, ..., max_coords = 5, precision = 9, trim = TRUE) {
-  paste0("<", s2_geography_format(x, max_coords, precision, trim), ">")
+  wk::wk_format(x, precision = precision, max_coords = max_coords, trim = trim)
 }
 
 # this is what gets called by the RStudio viewer, for which
